requirements.txt
pinecone>=2.0.0
requests>=2.30.0
python-dotenv>=1.0.0
transformers>=4.40.0
torch>=2.0.0
torchvision>=0.16.0
langgraph>=0.1.0
streamlit>=1.30.0
